Moved TS code into separate file, transpiler is now loaded in a background thread
